06_Modular_Self_RAG_Agent/requirement.txt
datasets
chromadb
ollama
google-genai